say something on restart
i think its funny. as an aside, maybe helga could use some sort of
instance specific code, as opposed to plugin

or maybe this should just stay in the plugin, i'll think about it. I
kinda intended this to be the only plugin for a particular bot, to allow
people to ignore it easily, or for pure isolation. its like the unix
philosophy for helga plugins.